refactor: (#281) 변경 적용 UseCase
