+++ simtong-application/src/main/kotlin/team/comit/simtong/domain/user/usecase/ChangeEmailUseCase.kt
@@ -2,7 +2,7 @@ package team.comit.simtong.domain.user.usecase
 
 import team.comit.simtong.domain.auth.exception.AuthExceptions
 import team.comit.simtong.domain.auth.spi.QueryAuthCodeLimitPort
-import team.comit.simtong.domain.user.dto.ChangeEmailRequest
+import team.comit.simtong.domain.user.dto.request.ChangeEmailData
 import team.comit.simtong.domain.user.exception.UserExceptions
 import team.comit.simtong.domain.user.spi.CommandUserPort
 import team.comit.simtong.domain.user.spi.QueryUserPort
@@ -25,7 +25,7 @@ class ChangeEmailUseCase(
     private val commandUserPort: CommandUserPort
 ) {
 
-    fun execute(request: ChangeEmailRequest) {
+    fun execute(request: ChangeEmailData) {
         if (queryUserPort.existsUserByEmail(request.email)) {
             throw AuthExceptions.AlreadyUsedEmail()
         }
