+++ simtong-application/src/main/kotlin/team/comit/simtong/domain/user/usecase/ChangePasswordUseCase.kt
@@ -1,6 +1,6 @@
 package team.comit.simtong.domain.user.usecase
 
-import team.comit.simtong.domain.user.dto.ChangePasswordRequest
+import team.comit.simtong.domain.user.dto.request.ChangePasswordData
 import team.comit.simtong.domain.user.exception.UserExceptions
 import team.comit.simtong.domain.user.spi.CommandUserPort
 import team.comit.simtong.domain.user.spi.QueryUserPort
@@ -22,7 +22,7 @@ class ChangePasswordUseCase(
     private val commandUserPort: CommandUserPort
 ) {
 
-    fun execute(request: ChangePasswordRequest) {
+    fun execute(request: ChangePasswordData) {
         val currentUserId = userSecurityPort.getCurrentUserId()
         val user = queryUserPort.queryUserById(currentUserId) ?: throw UserExceptions.NotFound()
