+++ simtong-application/src/main/kotlin/team/comit/simtong/domain/user/usecase/ChangeProfileImageUseCase.kt
@@ -2,7 +2,7 @@ package team.comit.simtong.domain.user.usecase
 
 import team.comit.simtong.domain.file.exception.FileExceptions
 import team.comit.simtong.domain.file.spi.CheckFilePort
-import team.comit.simtong.domain.user.dto.ChangeProfileImageRequest
+import team.comit.simtong.domain.user.dto.request.ChangeProfileImageData
 import team.comit.simtong.domain.user.exception.UserExceptions
 import team.comit.simtong.domain.user.spi.CommandUserPort
 import team.comit.simtong.domain.user.spi.QueryUserPort
@@ -25,7 +25,7 @@ class ChangeProfileImageUseCase(
     private val checkFilePort: CheckFilePort
 ) {
 
-    fun execute(request: ChangeProfileImageRequest) {
+    fun execute(request: ChangeProfileImageData) {
         if (!checkFilePort.existsPath(request.profileImagePath)) {
             throw FileExceptions.PathNotFound()
         }
